requirements.txt
numpy<2.0.0
fastapi
uvicorn
groq
pydantic>=2.0.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-groq
python-dotenv
pymupdf
python-multipart
langchain-huggingface
asyncpg==0.30.0
sqlalchemy[asyncio]>=2.0.0,<3.0.0
psycopg2-binary==2.9.10
e2b-code-interpreter>=1.0.0
# DOCBOT-201: credential encryption
cryptography>=41.0.0
# DOCBOT-203: AST-based SQL validation
sqlglot>=23.0.0
# DOCBOT-205: MySQL connector (not mysqlclient — build issues on Railway)
pymysql>=1.1.0
# DOCBOT-207: CSV parsing and type inference
pandas>=2.0.0
# DOCBOT-301: Python code generation via Qwen 2.5 Coder on Groq (uses existing groq package)
# DOCBOT-406: LangExtract for structured financial document extraction (Google, Gemini-backed)
langextract
# DOCBOT-208: Azure SQL / Microsoft Entra Service Principal auth
pyodbc>=5.0.0
azure-identity>=1.17.0
# DOCBOT-405: Analytical Autopilot — LangGraph state machine
langgraph>=0.2.0
# DOCBOT-601: SAML 2.0 SSO (Okta, Azure AD)
python3-saml>=1.16.0
# DOCBOT-701: Consumer OAuth (GitHub, Google) + email/password auth
httpx>=0.27.0
bcrypt>=4.0.0
# DOCBOT-1001: Chroma persistent vector store (replaces InMemoryVectorStore)
chromadb>=0.5.0
langchain-chroma>=0.1.0
# DOCBOT-1003: SemanticChunker for financial/legal document splitting
langchain-experimental>=0.0.50
# LLM Fallback: Gemini as fallback provider (direct SDK, avoids langchain-google-genai pydantic_v1 bug)
google-generativeai>=0.8.0
